# docs/接口文档/openapi打包.yaml
openapi: 3.0.3
info:
  title: Hero3 API
  version: 0.1.0
  description: Hero3 当前 MVP 阶段的后端接口文档，可导入 Apifox。
servers:
- url: http://localhost:8080
  description: 本地 Go 后端
- url: http://124.223.111.163:8080
  description: 服务器后端占位，部署后按实际端口调整
...
paths:
...
  /api/v1/game/state:
    get:
      tags:
      - Game
      summary: 获取游戏状态
      operationId: getGameState
      parameters:
      - name: playerId
        in: query
        required: true
        description: 玩家存档 ID。
        schema:
          type: string
          example: player_47f23e0eeb08ab50402fedfc
      responses:
        '200':
          description: 游戏状态快照
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameState'
        '404':
          description: 玩家存档不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: 缺少玩家存档 ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/accounts/register:
    post:
      tags:
      - Account
      summary: 注册轻账号
      operationId: registerAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSession'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 账号已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/accounts/login:
    post:
      tags:
      - Account
      summary: 登录轻账号
      operationId: loginAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSession'
        '401':
          description: 账号或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/accounts/{accountId}/players:
    get:
      tags:
      - Account
      summary: 查看账号绑定的存档
      operationId: listAccountPlayers
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: string
          example: acc_958538e3bc17faf00407ddcc
      responses:
        '200':
          description: 存档列表
          content:
            application/json:
              schema:
                type: object
                required:
                - players
                properties:
                  players:
                    type: array
                    items:
                      $ref: '#/components/schemas/PlayerSummary'
        '404':
          description: 账号不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/accounts/{accountId}:
    delete:
      tags:
      - Account
      summary: 删除账号
      description: 删除账号，同时删除该账号关联的全部云存档。
      operationId: deleteAccount
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: string
          example: acc_958538e3bc17faf00407ddcc
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '404':
          description: 账号不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/players/create:
    post:
      tags:
      - Account
      summary: 创建账号绑定的游戏存档
      operationId: createPlayer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlayerRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                required:
                - playerId
                - state
                properties:
                  playerId:
                    type: string
                    example: player_47f23e0eeb08ab50402fedfc
                  state:
                    $ref: '#/components/schemas/GameState'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 账号不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/players/{playerId}:
    delete:
      tags:
      - Account
      summary: 删除云存档
      description: 删除指定 playerId 对应的云存档，不删除账号。
      operationId: deletePlayer
      parameters:
      - name: playerId
        in: path
        required: true
        schema:
          type: string
          example: player_47f23e0eeb08ab50402fedfc
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '404':
          description: 存档不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/city/buildings/upgrade:
    post:
      tags:
      - City
      summary: 升级建筑
      description: 消耗资源发起建筑升级，升级完成前该建筑不可再次升级。
      operationId: upgradeBuilding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpgradeBuildingRequest'
      responses:
        '200':
          description: 升级已开始，返回更新后的游戏状态
          content:
            application/json:
              schema:
                type: object
                required:
                - state
                properties:
                  state:
                    $ref: '#/components/schemas/GameState'
        '404':
          description: 玩家或建筑不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 建筑正在升级中
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 资源不足或已满级
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/city/buildings/upgrade-batch:
    post:
      tags:
      - City
      summary: 批量升级资源建筑
      description: 按等级从低到高依次尝试升级所有空闲资源建筑，直到资源耗尽。
      operationId: upgradeBuildingBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - playerId
              properties:
                playerId:
                  type: string
                  example: player_47f23e0eeb08ab50402fedfc
      responses:
        '200':
          description: 批量升级完成
          content:
            application/json:
              schema:
                type: object
                required:
                - state
                - upgraded
                properties:
                  state:
                    $ref: '#/components/schemas/GameState'
                  upgraded:
                    type: integer
                    description: 本次成功升级的建筑数量
                    example: 3
        '404':
          description: 玩家不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 资源不足，无法升级任何建筑
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/city/resources/fill:
    post:
      tags:
      - City
      summary: GM 填满资源
      description: 将指定玩家的所有资源填满到仓库容量上限。
      operationId: fillResources
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - playerId
              properties:
                playerId:
                  type: string
                  example: player_47f23e0eeb08ab50402fedfc
      responses:
        '200':
          description: 资源已填满
          content:
            application/json:
              schema:
                type: object
                required:
                - state
                properties:
                  state:
                    $ref: '#/components/schemas/GameState'
        '404':
          description: 玩家不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/military/recruit:
    post:
      tags:
      - Military
      summary: 发起征兵
      description: 消耗资源创建征兵队列。队列到期后，后端在游戏状态结算时自动将兵力加入军队并移除该队列。
      operationId: recruitUnits
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecruitRequest'
      responses:
        '200':
          description: 征兵已进入队列，返回更新后的游戏状态
          content:
            application/json:
              schema:
                type: object
                required:
                - state
                properties:
                  state:
                    $ref: '#/components/schemas/GameState'
        '400':
          description: 征兵数量无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 玩家或兵种不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 征兵队列已满
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 资源不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/military/recruit/instant:
    post:
      tags:
      - Military
      summary: 极速完成征兵队列
      description: 立即完成指定征兵队列并重新计算后续队列结束时间。金币消耗暂未接入，当前用于开发和 GM 验证。
      operationId: instantCompleteRecruit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstantCompleteRecruitRequest'
      responses:
        '200':
          description: 征兵队列已完成，返回更新后的游戏状态
          content:
            application/json:
              schema:
                type: object
                required:
                - state
                properties:
                  state:
                    $ref: '#/components/schemas/GameState'
        '400':
          description: 参数错误或队列不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 玩家不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/military/general/stat:
    post:
      tags:
      - Military
      summary: 将领四维加点
      operationId: allocateGeneralStat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AllocateGeneralStatRequest'
      responses:
        '200':
          description: 加点成功，返回更新后的游戏状态
          content:
            application/json:
              schema:
                type: object
                required:
                - state
                properties:
                  state:
                    $ref: '#/components/schemas/GameState'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 玩家或将领不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 没有可用点数或属性已满
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/military/general/reset-stats:
    post:
      tags:
      - Military
      summary: 将领洗点
      description: 消耗账户金币 10 个，清空当前将领四维加点，保留将领、等级和经验，无冷却。
      operationId: resetGeneralStats
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetGeneralStatsRequest'
      responses:
        '200':
          description: 洗点成功，返回更新后的状态和账户金币余额
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralActionResult'
        '404':
          description: 玩家、账号或将领不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 金币不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/military/general/change:
    post:
      tags:
      - Military
      summary: 更换将领
      description: 仅允许更换当前存档阵营内的其它将领。保留等级和经验，重置四维加点。itemId 为预留换将卡消耗字段，传入时会消耗背包中 1
        个对应物品。
      operationId: changeGeneral
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeGeneralRequest'
      responses:
        '200':
          description: 换将成功，返回更新后的状态和账户金币余额
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralActionResult'
        '404':
          description: 玩家、将领或换将卡物品配置不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 将领不属于当前阵营、不能更换为当前将领，或换将卡库存不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/map/npc-cities:
    get:
      tags:
      - Map
      summary: 获取 NPC 城池列表
      description: 按玩家存档读取 NPC 城池状态。若该玩家尚未生成 NPC 城池，或已超过刷新周期，后端会生成并保存新的 NPC 状态。
      operationId: getNpcCities
      parameters:
      - name: playerId
        in: query
        required: true
        description: 玩家存档 ID。
        schema:
          type: string
          example: player_47f23e0eeb08ab50402fedfc
      responses:
        '200':
          description: NPC 城池状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NpcState'
        '404':
          description: 玩家不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/map/npc-cities/refresh:
    post:
      tags:
      - Map
      summary: 手动刷新 NPC 城池
      description: 重新生成该玩家的 NPC 城池列表。金币消耗暂未接入，当前用于开发和 GM 验证。
      operationId: refreshNpcCities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NpcCitiesRefreshRequest'
      responses:
        '200':
          description: 刷新后的 NPC 城池状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NpcState'
        '404':
          description: 玩家不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/map/npc-cities/attack:
    post:
      tags:
      - Map
      summary: 攻击或掠夺 NPC 城池
      description: 扣除出征兵力，执行战斗结算，归还存活兵力，更新 NPC 守军、NPC 资源、玩家资源和最近战报。
      operationId: attackNpcCity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttackNpcRequest'
      responses:
        '200':
          description: 战斗结算结果和更新后的游戏状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttackNpcResponse'
        '400':
          description: 参数错误或未选择出征兵力
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 玩家、NPC 城池或兵种不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 出征兵力不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/map/npc-cities/scout:
    post:
      tags:
      - Map
      summary: 侦查 NPC 城池
      description: 消耗侦察兵，返回目标 NPC 城池完整信息和更新后的游戏状态。
      operationId: scoutNpcCity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoutNpcRequest'
      responses:
        '200':
          description: 侦查结果和更新后的游戏状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoutNpcResponse'
        '400':
          description: 参数错误或没有可用侦察兵种
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 玩家或 NPC 城池不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 侦察兵数量不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/news/reports:
    get:
      tags:
      - News
      summary: 分页获取军情战报
      description: 获取玩家最近 3 天内未删除的军情战报。用于军情页分页展示，避免一次性拉取大量战报。
      operationId: listBattleReports
      parameters:
      - name: playerId
        in: query
        required: true
        description: 玩家存档 ID。
        schema:
          type: string
          example: player_47f23e0eeb08ab50402fedfc
      - name: page
        in: query
        required: false
        description: 页码，从 1 开始。
        schema:
          type: integer
          minimum: 1
          default: 1
      - name: pageSize
        in: query
        required: false
        description: 每页数量，最大 50。
        schema:
          type: integer
          minimum: 1
          maximum: 50
          default: 10
      responses:
        '200':
          description: 战报分页结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BattleReportPage'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权访问该玩家存档
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/mails:
    get:
      tags:
      - Mail
      summary: 分页获取信函
      parameters:
      - name: playerId
        in: query
        required: true
        schema:
          type: string
      - name: page
        in: query
        schema:
          type: integer
          minimum: 1
          default: 1
      - name: pageSize
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 50
          default: 10
      - name: mailType
        in: query
        description: 按信函类型筛选；不传或传 all 时返回全部。
        schema:
          type: string
          enum:
          - all
          - gm_notice
          - compensation
          - reward
          - event_reward
          - system_notice
          - player_message
      responses:
        '200':
          description: 信函分页结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MailPage'
  /api/v1/mails/{mailId}:
    get:
      tags:
      - Mail
      summary: 获取信函详情并标记已读
      parameters:
      - name: mailId
        in: path
        required: true
        schema:
          type: string
      - name: playerId
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: 信函详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mail'
        '404':
          description: 信函不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/mails/{mailId}/delete:
    post:
      tags:
      - Mail
      summary: 删除单封信函
      parameters:
      - name: mailId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - playerId
              properties:
                playerId:
                  type: string
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
  /api/v1/mails/{mailId}/claim:
    post:
      tags:
      - Mail
      summary: 领取信函附件
      parameters:
      - name: mailId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - playerId
              properties:
                playerId:
                  type: string
      responses:
        '200':
          description: 领取成功，返回信函和更新后的局部游戏状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MailClaimResult'
        '400':
          description: 无附件、已领取、已过期或玩家信函不可领取
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/mails/send-player:
    post:
      tags:
      - Mail
      summary: 玩家向玩家发送信函
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendPlayerMailRequest'
      responses:
        '201':
          description: 已创建的玩家信函
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mail'
        '400':
          description: 收件人格式错误、不能给自己发信或内容不合法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/minigame/record:
    post:
      tags:
      - MiniGame
      summary: 保存小游戏记录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MiniGameRecordRequest'
      responses:
        '201':
          description: 已创建记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MiniGameRecord'
...
  /api/v1/admin/mails/send:
    post:
      tags:
      - Admin
      summary: GM 给指定玩家发送信函
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMailRequest'
      responses:
        '201':
          description: 已创建的信函
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mail'
...
components:
  schemas:
...
    AccountRequest:
      type: object
      required:
      - username
      - password
      properties:
        username:
          type: string
          example: xieyuqi
        password:
          type: string
          format: password
          example: '123456'
    AccountSession:
      type: object
      required:
      - accountId
      - username
      properties:
        accountId:
          type: string
          example: acc_958538e3bc17faf00407ddcc
        username:
          type: string
          example: xieyuqi
...
    PlayerSummary:
      type: object
      required:
      - id
      - nickname
      - faction
      - updatedAt
      properties:
        id:
          type: string
          example: player_47f23e0eeb08ab50402fedfc
        nickname:
          type: string
          example: 主公
        mailCode:
          type: string
          description: 玩家信函短编号，与昵称组合成 玩家昵称#编号。
          example: '482193'
        faction:
          type: string
          example: wei
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: string
          example: player not found
    DeleteResponse:
      type: object
      required:
      - status
      properties:
        status:
          type: string
          example: deleted